\name{StephBoltz}
\alias{StephBoltz}
\title{
The Stephan Boltzman constant.
}
\description{
The Stephan Boltzman constant.  Units:  W/m^2/K^4 
}
\usage{
StephBoltz()
}

\author{
Glenn J Tattersall
}

\examples{

## The function is currently defined as
function () 
{
    s <- 5.67e-08
    s
  }
  
# Example
# This is simply the Stephan Boltzmann constant, saves having to remember the exact value
# and it allows easier coding.  To call it, type:

StephBoltz()

}

