\name{Prandtl}
\alias{Prandtl}
\title{
Returns the Prandtl number
}
\description{
Returns the Prandtl number
}
\usage{
Prandtl(Ta = 20)
}
\arguments{
  \item{Ta}{
Air temperature in degrees Celsius.  Default value is 20.
}
}
\details{
Returns the Prandlt number
}
\references{
Blaxter, K. 1989.  Energy Metabolism in Animals and Man
Gates, D. M. 2003.  Biophysical Ecology.  Dover Publications, Mineola, New York.  611 pp.
}
\author{
Glenn J Tattersall
}
\examples{

# Example:
Ta<-30
Prandtl(Ta)

}
