\name{Thermimage-package}
\alias{Thermimage-package}
\alias{Thermimage}
\docType{package}
\title{
Handles thermal image data input and conversion to temperature using established physical equations.
}
\description{
Assists in converting raw thermal imaging data files into temperature values.
}
\details{
\tabular{ll}{
Package: \tab Thermimage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-26\cr
License: \tab GPL-2\cr
}
Primary purpose of the package is to assist with manipulating raw data extracted from thermal image files.  These raw data are stored in a raw data format and require inforamtion about various environmental variables to estimate surface temperatures accurately.  raw2temp is the primary function of use.  Other functions included involve simple scripts for data handling.  
}
\author{
Glenn J. Tattersall <gtatters@brocku.ca>
}
\references{
1. http://130.15.24.88/exiftool/forum/index.php/topic,4898.60.html

2. Minkina, W. and Dudzik, S. 2009. Infrared Thermography: Errors and Uncertainties.  Wiley Press, 192 pp.
}
